Starting to implement the actual parsers.
For now, I just added the LiteralObjectParser. As well as the necessary protocols and types for ParseResults (both success and failures).
For the more primitive parsers I added a protocol ParserBuilder and I extended java.lang.Character with this protocol. I plan to do the same with strings and maps.